requirements.txt
fastapi
uvicorn[standard]
pydantic>=2
kafka-python